pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "benchlab"
version = "0.1.0"
description = "LLM Evaluation Engine - Run prompts against multiple LLMs, store results, and score with evaluator models"
readme = {text = "BenchLab - LLM Evaluation Engine", content-type = "text/plain"}
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "elasticsearch[async]>=8.10.0,<9",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
benchlab = "benchlab.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["benchlab"]
